+++ storage/transformers/transform_error.js
@@ -1,0 +1,6 @@
+export class TransformError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = "TransformError";
+  }
+}